feat: implement client feedback improvements for employee and admin APIs

+++ src/modules/users/users.controller.js
@@ -2,6 +2,7 @@
 
 const asyncHandler = require('../../common/asyncHandler');
 const { ok, noContent, created } = require('../../common/response');
+const ApiError = require('../../common/apiError');
 const pagination = require('../../common/pagination');
 const service = require('./users.service');
 
@@ -17,6 +18,8 @@ const getById = asyncHandler(async (req, res) => {
   return ok(res, data);
 });
 
+const levels = asyncHandler(async (_req, res) => ok(res, service.getLevels()));
+
 const create = asyncHandler(async (req, res) => {
   const data = await service.createUser(req.body);
   return created(res, data);
@@ -28,6 +31,9 @@ const update = asyncHandler(async (req, res) => {
 });
 
 const remove = asyncHandler(async (req, res) => {
+  if (Number(req.params.id) === req.user.id) {
+    throw ApiError.conflict('You cannot delete your own account');
+  }
   await service.removeUser(req.params.id);
   return noContent(res);
 });
@@ -35,6 +41,7 @@ const remove = asyncHandler(async (req, res) => {
 module.exports = {
   list,
   getById,
+  levels,
   create,
   update,
   remove,
